feature/12 refactor: update assets index - add asset: github logo
