--- src/assets/icons/index.ts
@@ -1,61 +1,0 @@
-import { FaArrowLeft, FaArrowRight } from 'react-icons/fa';
-import { IoMdArrowDropdown, IoMdClose, IoIosSearch } from 'react-icons/io';
-import { TfiReload } from 'react-icons/tfi';
-
-export {
-  FaArrowLeft as ArrowLeft,
-  FaArrowRight as ArrowRight,
-  IoMdArrowDropdown as ArrowDropdown,
-  TfiReload as Reload,
-  IoMdClose as Close,
-  IoIosSearch as Search,
-};
-
-export { default as blocksIcon } from './blocks.webp';
-export { default as cdPlayerIcon } from './cd_player.webp';
-export { default as closeIcon } from './close.webp';
-export { default as computerColorsIcon } from './computer_colors.webp';
-export { default as computerIcon } from './computer.webp';
-export { default as copyIcon } from './copy.webp';
-export { default as cutIcon } from './cut.webp';
-export { default as documentsIcon } from './documents.webp';
-export { default as domesticGroupIcon } from './domestic_group.webp';
-export { default as downloadsIcon } from './downloads.webp';
-export { default as excludeIcon } from './exclude.webp';
-export { default as favoritesIcon } from './favorites.webp';
-export { default as fileExplorerIcon } from './file_explorer.webp';
-export { default as flexBox1Icon } from './flex-box-1.webp';
-export { default as flexBox2Icon } from './flex-box-2.webp';
-export { default as folderIcon } from './folder-1.webp';
-export { default as folderUserIcon } from './folder-user.webp';
-export { default as galleryIcon } from './gallery.webp';
-export { default as gameHubIcon } from './game_hub.webp';
-export { default as imagesIcon } from './images.webp';
-export { default as indicationArrowIcon } from './indication_arrow.svg';
-export { default as internetExplorerExternalLinkIcon } from './internet_explorer_external_link.webp';
-export { default as internetExplorerIcon } from './internet_explorer.webp';
-export { default as locationsIcon } from './locations.webp';
-export { default as musicsIcon } from './musics.webp';
-export { default as networkIcon } from './network.webp';
-export { default as notepadIcon } from './notepad.webp';
-export { default as pasteIcon } from './paste.webp';
-export { default as personalUserIcon } from './personal_user.webp';
-export { default as pictureIcon } from './picture.webp';
-export { default as printerIcon } from './printer.webp';
-export { default as questionIcon } from './question.webp';
-export { default as saveIcon } from './save.webp';
-export { default as sheetMusicIcon } from './sheet_music.webp';
-export { default as usersIcon } from './users.webp';
-export { default as videosIcon } from './videos.webp';
-export { default as workspaceIcon } from './workspace.webp';
-
-export { default as mediaCenterCloseIcon } from './temp/close.webp';
-export { default as mediaCenterExpandIcon } from './temp/expand.webp';
-export { default as mediaCenterImageBtnIcon } from './temp/image_btn.webp';
-export { default as mediaCenterNextIcon } from './temp/next.webp';
-export { default as mediaCenterPreviousIcon } from './temp/previous.webp';
-export { default as mediaCenterRedoIcon } from './temp/redo.webp';
-export { default as mediaCenterUndoIcon } from './temp/undo.webp';
-export { default as mediaCenterZoomLess } from './temp/zoom_less.webp';
-export { default as mediaCenterZoomMore } from './temp/zoom_more.webp';
-export { default as mediaCenterImageIcon } from './temp/media_center_image.webp';
