+++ src/assets/index.ts
@@ -1,0 +1,68 @@
+import { FaArrowLeft, FaArrowRight } from 'react-icons/fa';
+import { IoMdArrowDropdown, IoMdClose, IoIosSearch } from 'react-icons/io';
+import { TfiReload } from 'react-icons/tfi';
+
+// NOTE: react-icons
+export {
+  FaArrowLeft as ArrowLeft,
+  FaArrowRight as ArrowRight,
+  IoMdArrowDropdown as ArrowDropdown,
+  TfiReload as Reload,
+  IoMdClose as Close,
+  IoIosSearch as Search,
+};
+
+// NOTE: windows icons
+export { default as blocksIcon } from './icons/blocks.webp';
+export { default as cdPlayerIcon } from './icons/cd_player.webp';
+export { default as closeIcon } from './icons/close.webp';
+export { default as computerColorsIcon } from './icons/computer_colors.webp';
+export { default as computerIcon } from './icons/computer.webp';
+export { default as copyIcon } from './icons/copy.webp';
+export { default as cutIcon } from './icons/cut.webp';
+export { default as documentsIcon } from './icons/documents.webp';
+export { default as domesticGroupIcon } from './icons/domestic_group.webp';
+export { default as downloadsIcon } from './icons/downloads.webp';
+export { default as excludeIcon } from './icons/exclude.webp';
+export { default as favoritesIcon } from './icons/favorites.webp';
+export { default as fileExplorerIcon } from './icons/file_explorer.webp';
+export { default as flexBox1Icon } from './icons/flex-box-1.webp';
+export { default as flexBox2Icon } from './icons/flex-box-2.webp';
+export { default as folderIcon } from './icons/folder-1.webp';
+export { default as folderUserIcon } from './icons/folder-user.webp';
+export { default as galleryIcon } from './icons/gallery.webp';
+export { default as gameHubIcon } from './icons/game_hub.webp';
+export { default as imagesIcon } from './icons/images.webp';
+export { default as indicationArrowIcon } from './icons/indication_arrow.svg';
+export { default as internetExplorerExternalLinkIcon } from './icons/internet_explorer_external_link.webp';
+export { default as internetExplorerIcon } from './icons/internet_explorer.webp';
+export { default as locationsIcon } from './icons/locations.webp';
+export { default as musicsIcon } from './icons/musics.webp';
+export { default as networkIcon } from './icons/network.webp';
+export { default as notepadIcon } from './icons/notepad.webp';
+export { default as pasteIcon } from './icons/paste.webp';
+export { default as personalUserIcon } from './icons/personal_user.webp';
+export { default as pictureIcon } from './icons/picture.webp';
+export { default as printerIcon } from './icons/printer.webp';
+export { default as questionIcon } from './icons/question.webp';
+export { default as saveIcon } from './icons/save.webp';
+export { default as sheetMusicIcon } from './icons/sheet_music.webp';
+export { default as usersIcon } from './icons/users.webp';
+export { default as videosIcon } from './icons/videos.webp';
+export { default as workspaceIcon } from './icons/workspace.webp';
+
+// NOTE: media center icons
+export { default as mediaCenterCloseIcon } from './icons/temp/close.webp';
+export { default as mediaCenterExpandIcon } from './icons/temp/expand.webp';
+export { default as mediaCenterImageBtnIcon } from './icons/temp/image_btn.webp';
+export { default as mediaCenterNextIcon } from './icons/temp/next.webp';
+export { default as mediaCenterPreviousIcon } from './icons/temp/previous.webp';
+export { default as mediaCenterRedoIcon } from './icons/temp/redo.webp';
+export { default as mediaCenterUndoIcon } from './icons/temp/undo.webp';
+export { default as mediaCenterZoomLess } from './icons/temp/zoom_less.webp';
+export { default as mediaCenterZoomMore } from './icons/temp/zoom_more.webp';
+export { default as mediaCenterImageIcon } from './icons/temp/media_center_image.webp';
+
+// NOTE: Logos
+export { default as GithubLogo } from './logo/github_logo.webp';
+export { default as WindowsLogo } from './logo/windows-logo.png';
