+++ src/components/apps/window/window-layout.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { closeIcon } from '@/assets/icons';
+import { closeIcon } from '@/assets';
 
 interface WindowLayoutProps {
   children: React.ReactNode;
