+++ src/components/common/buttons/btn-forward-backward/btn-forward-backward.tsx
@@ -1,4 +1,4 @@
-import { ArrowDropdown, ArrowLeft, ArrowRight } from '@/assets/icons';
+import { ArrowDropdown, ArrowLeft, ArrowRight } from '@/assets';
 import './btn-forward-backward.scss';
 
 interface BtnForwardBackwardProps {
